Feat: big int quantity fields - 2668

+++ backend/lcfs/db/models/transfer/Transfer.py
@@ -1,6 +1,6 @@
 import enum
 
-from sqlalchemy import Column, Integer, ForeignKey, DateTime, Enum, String, Numeric
+from sqlalchemy import Column, Integer, ForeignKey, DateTime, Enum, String, Numeric, BigInteger
 from sqlalchemy.orm import relationship
 
 from lcfs.db.base import BaseModel, Auditable, EffectiveDates
@@ -33,7 +33,7 @@ class Transfer(BaseModel, Auditable, EffectiveDates):
         Numeric(10, 2),
         comment="Price per unit with two decimal places",
     )
-    quantity = Column(Integer, comment="Quantity of units")
+    quantity = Column(BigInteger, comment="Quantity of units")
     transfer_category_id = Column(
         Integer, ForeignKey("transfer_category.transfer_category_id")
     )
